debug

+++ 2017/dnn_rl.py
@@ -135,8 +135,8 @@ class MMSE_loss(torch.nn.Module):
 
         #Generate predicted q_func
 
-        q_pred = torch.tensor(x_out.reshape(-1, 1)).cuda()
-        q_target = torch.tensor(A_t).cuda()
+        q_pred = torch.tensor(x_out.reshape(-1, 1)).cuda().float()
+        q_target = torch.tensor(A_t).cuda().float()
         print('NN out:', q_pred.size())
         print('Q-pred:', q_target.size())
